Add populateList.js file
