+++ webpack.config.js
@@ -3,7 +3,7 @@ const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
   mode: 'development',
-  entry: './src/index.js',
+  entry: ['./src/index.js', './src/populateList.js'],
   output: {
     filename: 'bundle.js',
     path: path.resolve(__dirname, 'dist'),
